Add batch processing and postgresql support (#4)
* add postgresql support

* add link and segment batch processing

* fix working dir write test concurrency

* improv hierarchy memory usage

* switch from tracks creation use from queue to list

* add numba track id assignment implementation

* add multiprocessing to generic export

* fix sql solution submission

* add poetry install workflow

* improving track id computation

* fix minor bugs

* add testing.postgresql tox dep

* fix overwrite when batch processing

* add postgresql windows test skip

+++ ultrack/core/segmentation/hierarchy.py
@@ -54,7 +54,7 @@ def create_hierarchies(
         or binary_detection.dtype == bool
     )
 
-    labels = measure.label(binary_detection, connectivity=1)
+    labels = measure.label(binary_detection, connectivity=1).astype(np.int32)
 
     if "min_area" in kwargs:
         labels = morphology.remove_small_objects(labels, min_size=kwargs["min_area"])
